% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_samsung.R
\name{add_prefix_to_filename}
\alias{add_prefix_to_filename}
\title{Add appropriate prefix to file names}
\usage{
add_prefix_to_filename(filename)
}
\arguments{
\item{filename}{Character string of the filename}
}
\value{
Character string of the processed filename
}
\description{
This function adds "IMG_" to image files and "VID_" to video files
that start with 8 digits.
}
