% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_samsung.R
\name{rename_samsung}
\alias{rename_samsung}
\title{Process and copy files with prefixes to a destination folder}
\usage{
rename_samsung(path)
}
\arguments{
\item{path}{Character string of the source directory path}
}
\value{
Invisible list of processed files
}
\description{
This function processes image and video files by adding the appropriate prefix
to those starting with 8 digits and copies them to a "renamed_pictures" folder.
}
