% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_samsung_from_whatsapp.R
\name{rename_samsung_from_whatsapp}
\alias{rename_samsung_from_whatsapp}
\title{Remove WhatsApp-style prefix from file names}
\usage{
rename_samsung_from_whatsapp(path)
}
\arguments{
\item{path}{Character string of the source directory path}
}
\value{
Invisible list of processed files
}
\description{
This function removes "IMG-" or "VID-" from the start of file names
and copies them to a "renamed_pictures" folder.
}
